feat: Modificado tipagem de saldo;
Double -> Float

+++ src/main/java/com/xpcosmos/transacoes_bancarias/dto/UserDTO.java
@@ -8,6 +8,6 @@ public record UserDTO(
     String documentoId,
     String email,
 		UserType tipoUsuario,
-		Double saldo,
+		Float saldo,
     String senha) {
 }
